add big boxes

+++ Tap Box/Assets/Scripts/LevelCreator/BigBoxPart.cs
@@ -7,9 +7,8 @@ public class BigBoxPart : MonoBehaviour
 
     public void UpdateArrayPosition()
     {
-        ArrayPosition = new Vector3((int)(transform.position.x / _size),
-            (int)(transform.position.y / _size),
-            (int)(transform.position.z / _size));
-
+        ArrayPosition = new Vector3((transform.position.x / _size),
+            (transform.position.y / _size),
+            (transform.position.z / _size));
     }
 }
